# docs/swagger.yaml
consumes:
- application/json
definitions:
  adminTool.IdRequest:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  adminTool.LeagueAddRequest:
    properties:
      name:
        type: string
      sport_id:
        type: integer
    required:
    - sport_id
    type: object
  adminTool.LeagueUpdateRequest:
    properties:
      id:
        type: integer
      name:
        type: string
      sport_id:
        type: integer
    required:
    - id
    - sport_id
    type: object
  adminTool.Sports:
    properties:
      id:
        type: integer
      name:
        type: string
    required:
    - id
    - name
    type: object
  adminTool.SportsAddRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  adminTool.SportsUpdateRequest:
    properties:
      id:
        type: integer
      name:
        type: string
    required:
    - id
    - name
    type: object
  adminTool.StadiumAddRequest:
    properties:
      address:
        type: string
      capacity:
        type: integer
      description:
        type: string
      fileId:
        type: string
      image:
        type: string
      name:
        type: string
    required:
    - address
    - capacity
    - description
    - fileId
    - image
    - name
    type: object
  adminTool.StadiumUpdateRequest:
    properties:
      address:
        type: string
      capacity:
        type: integer
      description:
        type: string
      fileId:
        type: string
      id:
        type: integer
      image:
        type: string
      name:
        type: string
    required:
    - address
    - capacity
    - description
    - fileId
    - id
    - image
    - name
    type: object
  adminTool.TeamAddRequest:
    properties:
      LeagueId:
        type: integer
      name:
        type: string
      sportsId:
        type: integer
      stadiumId:
        type: integer
    required:
    - LeagueId
    - name
    - sportsId
    - stadiumId
    type: object
  adminTool.TeamUpdateRequest:
    properties:
      LeagueId:
        type: integer
      name:
        type: string
      sportsId:
        type: integer
      stadiumId:
        type: integer
      teamId:
        type: integer
    required:
    - LeagueId
    - name
    - sportsId
    - stadiumId
    - teamId
    type: object
...
  utils.ApiResponse-adminTool_Sports:
    properties:
      data:
        $ref: '#/definitions/adminTool.Sports'
      message:
        example: 成功しました！！
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  utils.ErrorBasicResponse:
    properties:
      message:
        example: エラーメッセージ
        type: string
      success:
        example: false
        type: boolean
    type: object
  utils.SuccessBasicResponse:
    properties:
      message:
        example: 成功しました！！
        type: string
      success:
        example: true
        type: boolean
    type: object
info:
  contact: {}
  description: このapiは、ビジターゴーのAPIで、ユーザー、スタジアム、遠征記録、などについて扱います。
  title: ビジターゴーAPI
  version: "1.0"
paths:
...
  /api/admin/league/delete:
    delete:
      description: リクエストボディに削除対象のIDを指定してリーグ情報を削除します
      parameters:
      - description: リーグID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/adminTool.IdRequest'
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.SuccessBasicResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      security:
      - BearerAuth: []
      summary: リーグ削除
      tags:
      - league
  /api/admin/league/leagueAdd:
    post:
      description: リクエストからリーグ情報を取得後、重複確認を行い登録する。
      parameters:
      - description: リーグ情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/adminTool.LeagueAddRequest'
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.SuccessBasicResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      security:
      - BearerAuth: []
      summary: リーグの追加
      tags:
      - league
  /api/admin/league/leagues:
    get:
      description: リーグ情報のレコードを全件取得して、一覧として表示する。
      parameters:
      - description: キーワード
        in: query
        name: keyword
        type: string
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.SuccessBasicResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      summary: リーグ全件検索
      tags:
      - league
  /api/admin/league/update:
    put:
      description: リクエストボディに更新対象のIDを指定してリーグ情報を更新します
      parameters:
      - description: リーグ情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/adminTool.LeagueUpdateRequest'
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.SuccessBasicResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      security:
      - BearerAuth: []
      summary: リーグ更新
      tags:
      - league
...
  /api/admin/sports/delete/{id}:
    delete:
      description: リクエストボディに削除対象のIDを指定してスポーツ情報を削除します
      parameters:
      - description: スポーツID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.SuccessBasicResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      security:
      - BearerAuth: []
      summary: スポーツ削除
      tags:
      - sports
  /api/admin/sports/idSports/{id}:
    get:
      description: idからスポーツ情報のレコードを取得して表示する。
      parameters:
      - description: スポーツID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.ApiResponse-adminTool_Sports'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      summary: スポーツid検索
      tags:
      - sports
  /api/admin/sports/sports:
    get:
      description: スポーツ情報のレコードを全件取得して、一覧として表示する。
      parameters:
      - description: キーワード
        in: query
        name: keyword
        type: string
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.SuccessBasicResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      summary: スポーツ全件検索
      tags:
      - sports
  /api/admin/sports/sportsAdd:
    post:
      description: リクエストからスポーツ情報を取得後、重複確認を行い登録する。
      parameters:
      - description: スポーツ情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/adminTool.SportsAddRequest'
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.SuccessBasicResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      security:
      - BearerAuth: []
      summary: スポーツの追加
      tags:
      - sports
  /api/admin/sports/update:
    put:
      description: リクエストボディに更新対象のIDを指定してスポーツ情報を更新します
      parameters:
      - description: スポーツ情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/adminTool.SportsUpdateRequest'
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.SuccessBasicResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      security:
      - BearerAuth: []
      summary: スポーツ更新
      tags:
      - sports
  /api/admin/stadium/delete:
    delete:
      description: リクエストボディに削除対象のIDを指定してスタジアムを削除します
      parameters:
      - description: スタジアムID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/adminTool.IdRequest'
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.SuccessBasicResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      security:
      - BearerAuth: []
      summary: スタジアム削除
      tags:
      - stadium
  /api/admin/stadium/stadiumAdd:
    post:
      description: リクエストからスタジアム情報を追加後、重複確認を行い登録する。
      parameters:
      - description: スタジアム情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/adminTool.StadiumAddRequest'
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.SuccessBasicResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      security:
      - BearerAuth: []
      summary: スタジアム追加
      tags:
      - stadium
  /api/admin/stadium/stadiums:
    get:
      description: スタジアム情報のレコードを全件取得して、一覧として表示する。
      parameters:
      - description: キーワード
        in: query
        name: keyword
        type: string
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.SuccessBasicResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      summary: スタジアム全件検索
      tags:
      - stadium
  /api/admin/stadium/update:
    put:
      description: リクエストボディに更新対象のIDを指定してスタジアムを更新します
      parameters:
      - description: スタジアム情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/adminTool.StadiumUpdateRequest'
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.SuccessBasicResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      security:
      - BearerAuth: []
      summary: スタジアム更新
      tags:
      - stadium
  /api/admin/team/delete:
    delete:
      description: リクエストボディに削除対象のIDを指定してチーム情報を削除します
      parameters:
      - description: チームID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/adminTool.IdRequest'
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.SuccessBasicResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      security:
      - BearerAuth: []
      summary: チーム削除
      tags:
      - team
  /api/admin/team/teamAdd:
    post:
      description: リクエストからチーム情報を取得後、重複確認を行い登録する。
      parameters:
      - description: チーム情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/adminTool.TeamAddRequest'
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.SuccessBasicResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      security:
      - BearerAuth: []
      summary: チームの追加
      tags:
      - team
  /api/admin/team/teams:
    get:
      description: チーム情報のレコードを全件取得して、一覧として表示する。
      parameters:
      - description: キーワード
        in: query
        name: keyword
        type: string
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.SuccessBasicResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      summary: チーム全件検索
      tags:
      - team
  /api/admin/team/update:
    put:
      description: リクエストボディに更新対象のIDを指定してチーム情報を更新します
      parameters:
      - description: チーム情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/adminTool.TeamUpdateRequest'
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.SuccessBasicResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      security:
      - BearerAuth: []
      summary: チーム情報更新
      tags:
      - team
...
produces:
- application/json
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"